Use defer instead of async for script tags
Is there any reason to use `async` instead of `defer` here?

Scripts marked defer are executed (after parsing completes) in the order which they are defined in the markup.
Maybe in your case it doesn't matter what to use, but it could be helpful when i.e. you want to load some polyfill before other scripts, you mark it as defer not to block page rendering, but you cannot be sure it will be executed before other scripts, because they are marked async, and they can be executed before the deferred one.